Refactor - generateTargetNumbers.js 제거 및 내부 함수 이동 - 게임 진행시 필요한 함수의 규모가 크지 않아 gameUtil이라는 파일에 통합 - gameUtil 파일 내부에 정답 숫자 생성, 카운트 계산 함수 작성
