+++ src/utils/gameUtil.js
@@ -1,3 +1,4 @@
+// 스트라이크, 볼 카운트에 대한 Hint 관련
 export const calculateCount = (target, inputs) => {
   const strike = target.filter((num, i) => num === inputs[i]).length;
   const ball = target
@@ -19,4 +20,19 @@ export const getHint = ({ strike, ball }) => {
   result = result.trim();
 
   return result ? result : '낫씽';
+};
+
+// 정답 숫자 생성 관련
+const generateNumber = () => {
+  return Math.floor(Math.random() * 9) + 1;
+};
+
+export const generateTargetNumbers = () => {
+  const result = new Set();
+
+  while (result.size !== 3) {
+    result.add(generateNumber());
+  }
+
+  return [...result];
 };
